Moved the Q function context constructors back to {swe_,sediment}_ceed.c.
This makes sense, since these types are closely related to Q functions.
