+++ src/CMakeLists.txt
@@ -12,7 +12,9 @@ set(RDYCORE_C_SOURCE_FILES
   operator.c
   operator_ceed.c
   operator_petsc.c
+  swe/swe_ceed.c
   swe/swe_petsc.c
+  sediment/sediment_ceed.c
   sediment/sediment_petsc.c
   time_series.c
   xdmf_output.c
